requirements.txt
accelerate==0.15.0
datasets==2.8.0
deepspeed==0.7.7
einops==0.6.0
faiss-gpu==1.7.2
graphviz==0.16
joblib==1.2.0
nltk==3.8
numpy==1.24.0
pandas==1.5.2
pomegranate==0.14.8
python-decouple==3.6
PyYAML==6.0
scikit-learn==1.2.0
scipy==1.9.3
spacy==3.4.4
tensorboard==2.11.0
tensorflow==2.11.0
tensorflow-gpu==2.11.0
tokenizers==0.12.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
transformers==4.21.2
wandb==0.13.7